requirements.txt
eth_account==0.5.9
msgpack_python==0.5.6
py_solc_x==1.1.1
requests==2.28.1
web3==5.31.3
pynacl==1.5.0
coincurve==18.0.0
eip712-structs==1.1.0